greutate peste

+++ models/cart-item.js
@@ -10,7 +10,7 @@ const CartItem = sequelize.define('cartItem', {
         allowNull: false,
         primaryKey: true
     },
-    quantity: Sequelize.INTEGER
+    quantity: Sequelize.FLOAT
 });
 
 module.exports = CartItem;
